Adding type alias for borrowed_subrange_t
